+++ resources/compiledJS/script/ability.js
@@ -298,7 +298,8 @@ function getAbiStatusModifiers(char, abilityId, effectId) {
                     key = key.replace(abilityId + "_", "");
                     if (key.includes("status_effect")) {
                         const _key = key.replace("status_effect_", "");
-                        const __key = _key.substring(0, _key.length - 1);
+                        const trueKey = _key.replace(effectId + "_", "");
+                        const __key = trueKey.substring(0, trueKey.length - 1);
                         if (possible_stat_modifiers.find((m) => m == __key.toString())) {
                             if (key.endsWith("V"))
                                 total["effects"][__key].value += value;