+++ resources/compiledJS/script/localisation.js
@@ -14,6 +14,7 @@ const finnish = {
     core_stats: "Yleiset statsit",
     core_resistances: "puolustukset",
     status_resistances: "vastustuskyky",
+    passives: "Passiiviset Kyvyt",
     poison_status: "Myrkky",
     burning_status: "Palaminen",
     curse_status: "Kirous",
@@ -405,9 +406,12 @@ const finnish = {
     buy_item: "Osta esine",
     sell_item: "Myy esine",
     quest_added: "<c>goldenrod<c>Tehtävä '<c>gold<c>[QUEST]<c>goldenrod<c>' alkanut!",
+    revive: "Olet herännyt henkiin pyhäköllä!",
     // Character names
     testMerchant_name: "Karl Bauer",
     // Passive abilities
+    resilience_of_the_lone_wanderer_name: "Kulkijan Sinnikkyys",
+    lone_wanderer_desc: "Edes tuonela ei kykene pysäyttämään tehtäväsi suoritusta.",
     frenzy_name: "Kiihko",
     heightened_casting_name: "Vahva Alku",
     warrior_instinct_name: "Soturin Aisti",
@@ -420,6 +424,10 @@ const finnish = {
     sense_of_danger_2_name: "Kuudes Aisti II",
     frantic_mana_recovery_name: "Hätäinen Taian Palautus",
     mark_of_hunter_name: "Suojelijan Merkki",
+    racial_ability_orc_1_name: "Örkin Lihaksikkuus",
+    racial_ability_human_1_name: "Ihmisen Kestävyys",
+    racial_ability_elf_1_name: "Haltian Viisaus",
+    racial_ability_ashen_1_name: "Kalpean Ketteryys",
     // Enemy types
     against_type_syntax: " vastaan",
     singular_type_slime: "Lima",
@@ -854,6 +862,7 @@ const english = {
     core_stats: "Core stats",
     core_resistances: "Core resistances",
     status_resistances: "Status Resists",
+    passives: "Passive Abilities",
     poison_status: "Poison",
     burning_status: "Burning",
     curse_status: "Curse",
@@ -1142,6 +1151,7 @@ const english = {
     buy_item: "Buy item",
     sell_item: "Sell item",
     quest_added: "<c>goldenrod<c>Quest '<c>gold<c>[QUEST]<c>goldenrod<c>' has begun!",
+    revive: "You have been revived at a shrine!",
     // Character names
     testMerchant_name: "Karl Bauer",
     // Artifacts
@@ -1156,6 +1166,8 @@ const english = {
     warriorSet_name: "Warrior",
     loneShadeSet_name: "Lone Shade",
     // Passive abilities
+    resilience_of_the_lone_wanderer_name: "Lone Wanderer's Resilience",
+    lone_wanderer_desc: "You have a duty to complete, and nothing will stop you, not even death itself.",
     frenzy_name: "Frenzy",
     heightened_casting_name: "Heightened Casting",
     warrior_instinct_name: "Warrior Instinct",
@@ -1168,6 +1180,10 @@ const english = {
     sense_of_danger_2_name: "Danger Sense II",
     frantic_mana_recovery_name: "Frantic Mana Recovery",
     mark_of_hunter_name: "Mark of the Hunter",
+    racial_ability_orc_1_name: "Orcish Musculature",
+    racial_ability_human_1_name: "Human Tenacity",
+    racial_ability_elf_1_name: "Elven Wisdom",
+    racial_ability_ashen_1_name: "Ashen Agility",
     // Enemy types
     against_type_syntax: " against ",
     singular_type_slime: "Slime",