+++ resources/compiledJS/script/perk.js
@@ -322,6 +322,8 @@ function statModifTT(statModif) {
     var _a;
     statModif = new PermanentStatModifier(Object.assign({}, statModif));
     let txt = `§\n ${lang["passive"]} <f>16px<f><c>white<c>'<c>gold<c>${(_a = lang[statModif.id + "_name"]) !== null && _a !== void 0 ? _a : statModif.id}<c>white<c>'\n`;
+    if (statModif.desc)
+        txt += `§<c>silver<c><f>13px<f>"${lang[statModif.desc]}"\n§`;
     if (statModif.conditions) {
         txt += `<c>white<c><f>15px<f>${lang["active_if"]}:\n`;
         Object.entries(statModif.conditions).forEach(cond => {