+++ resources/compiledJS/script/player.js
@@ -259,6 +259,11 @@ class PlayerCharacter extends Character {
                 displayText(`<c>white<c>[WORLD] <c>gold<c>${lvlText}`);
                 updateUI();
             }
+            // Add racial bonus
+            if (this.level.level >= 10 && this.statModifiers.findIndex((m) => m.id === `racial_ability_${this.race}_1`) === -1) {
+                this.statModifiers.push(new PermanentStatModifier(statModifiers[`racial_ability_${this.race}_1`]));
+                spawnFloatingText(this.cords, "RACIAL ABILITY!", "lime", 50, 2000, 450);
+            }
         };
         this.getArtifactSetBonuses = (getOnlySetAmounts = false) => {
             var _a, _b, _c;
@@ -598,7 +603,7 @@ var player = new PlayerCharacter({
     },
     sprite: ".player",
     race: "orc",
-    hair: 3,
+    hair: 5,
     eyes: 2,
     face: 1,
     weapon: {},
@@ -642,7 +647,7 @@ var player = new PlayerCharacter({
     grave: null,
     flags: {},
     questProgress: [],
-    sex: "male"
+    sex: "female"
 });
 let combatSummons = [];
 var randomProperty = function (mods) {
@@ -697,7 +702,7 @@ function respawnPlayer() {
     updateUI();
     resetAllLivingEnemiesInAllMaps();
     modifyCanvas(true);
-    displayText("HERÄSIT KUOLLEISTA!");
+    displayText(`[WORLD] ${lang["revive"]}`);
     spawnFloatingText(player.cords, "REVIVE!", "green", 36, 575, 75);
 }
 // for (let i = 0; i < 20; i++) {