+++ resources/compiledJS/script/quest.js
@@ -184,7 +184,9 @@ function spawnQuestMonsters() {
                         foundUniqueMob = true;
                 });
                 if (!foundUniqueMob) {
-                    maps.find((m) => m.id == enemy.map).enemies.push(new Enemy(Object.assign(Object.assign({}, enemies[enemy.enemy]), { cords: enemy.pos, spawnCords: enemy.pos, level: enemy.level, map: enemy.map, questSpawn: { quest: q.id, index: index } })));
+                    let spawnMap = maps.find((m) => m.id == enemy.map);
+                    spawnMap.enemies.push(new Enemy(Object.assign(Object.assign({}, enemies[enemy.enemy]), { cords: enemy.pos, spawnCords: enemy.pos, level: enemy.level, map: enemy.map, questSpawn: { quest: q.id, index: index } })));
+                    spawnMap.enemies[spawnMap.enemies.length - 1].updateStatModifiers();
                 }
             });
         }