+++ script/map.ts
@@ -824,9 +824,10 @@ function canMoveTo(char: any, tile: tileObject) {
   return movable;
 }
 
-function renderPlayerOutOfMap(size: number, canvas: HTMLCanvasElement, ctx: any, side: string = "center", playerModel: any = player) {
+function renderPlayerOutOfMap(size: number, canvas: HTMLCanvasElement, ctx: any, side: string = "center", playerModel: any = player, noClothes: boolean = false) {
   canvas.width = canvas.width; // Clear canvas
-  const bodyModel = <HTMLImageElement>document.querySelector(".sprites ." + playerModel.race + "Model" + capitalizeFirstLetter(player.sex));
+  const sex = playerModel.sex === "male" ? "" : capitalizeFirstLetter(playerModel.sex);
+  const bodyModel = <HTMLImageElement>document.querySelector(".sprites ." + playerModel.race + "Model" + capitalizeFirstLetter(playerModel.sex));
   const earModel = <HTMLImageElement>document.querySelector(".sprites ." + playerModel.race + "Ears");
   const hairModel = <HTMLImageElement>document.querySelector(".sprites .hair" + playerModel.hair);
   const eyeModel = <HTMLImageElement>document.querySelector(".sprites .eyes" + playerModel.eyes);
@@ -840,30 +841,32 @@ function renderPlayerOutOfMap(size: number, canvas: HTMLCanvasElement, ctx: any,
   ctx?.drawImage(eyeModel, x, y, size, size);
   ctx?.drawImage(faceModel, x, y, size, size);
   ctx?.drawImage(leggings, x, y, size, size);
-  if (!playerModel.helmet?.coversHair) ctx?.drawImage(hairModel, x, y, size, size);
-  if (playerModel.helmet?.sprite) {
-    const helmetModel = <HTMLImageElement>document.querySelector(".sprites ." + playerModel.helmet.sprite);
-    ctx?.drawImage(helmetModel, x, y, size, size);
-  }
-  if (playerModel.gloves?.sprite) {
-    const glovesModel = <HTMLImageElement>document.querySelector(".sprites ." + playerModel.gloves.sprite);
-    ctx?.drawImage(glovesModel, x, y, size, size);
-  }
-  if (playerModel.boots?.sprite) {
-    const bootsModel = <HTMLImageElement>document.querySelector(".sprites ." + playerModel.boots.sprite);
-    ctx?.drawImage(bootsModel, x, y, size, size);
-  }
-  if (playerModel.legs?.sprite) {
-    const leggingsModel = <HTMLImageElement>document.querySelector(".sprites ." + playerModel.legs.sprite);
-    ctx?.drawImage(leggingsModel, x, y, size, size);
-  }
-  else if (!playerModel.legs?.sprite) {
-    const leggings = <HTMLImageElement>document.querySelector(`.sprites .defaultPants${capitalizeFirstLetter(player.sex)}`);
-    ctx?.drawImage(leggings, x, y, size, size);
-  }
-  if (playerModel.chest?.sprite) {
-    const chestModel = <HTMLImageElement>document.querySelector(".sprites ." + playerModel.chest.sprite);
-    ctx?.drawImage(chestModel, x, y, size, size);
+  if (!playerModel.helmet?.coversHair || noClothes) ctx?.drawImage(hairModel, x, y, size, size);
+  if (!noClothes) {
+    if (playerModel.helmet?.sprite) {
+      const helmetModel = <HTMLImageElement>document.querySelector(".sprites ." + playerModel.helmet.sprite + sex);
+      ctx?.drawImage(helmetModel, x, y, size, size);
+    }
+    if (playerModel.gloves?.sprite) {
+      const glovesModel = <HTMLImageElement>document.querySelector(".sprites ." + playerModel.gloves.sprite + sex);
+      ctx?.drawImage(glovesModel, x, y, size, size);
+    }
+    if (playerModel.boots?.sprite) {
+      const bootsModel = <HTMLImageElement>document.querySelector(".sprites ." + playerModel.boots.sprite + sex);
+      ctx?.drawImage(bootsModel, x, y, size, size);
+    }
+    if (playerModel.legs?.sprite) {
+      const leggingsModel = <HTMLImageElement>document.querySelector(".sprites ." + playerModel.legs.sprite + sex);
+      ctx?.drawImage(leggingsModel, x, y, size, size);
+    }
+    else if (!playerModel.legs?.sprite) {
+      const leggings = <HTMLImageElement>document.querySelector(`.sprites .defaultPants${capitalizeFirstLetter(player.sex)}`);
+      ctx?.drawImage(leggings, x, y, size, size);
+    }
+    if (playerModel.chest?.sprite) {
+      const chestModel = <HTMLImageElement>document.querySelector(".sprites ." + playerModel.chest.sprite + sex);
+      ctx?.drawImage(chestModel, x, y, size, size);
+    }
   }
   if (playerModel.weapon?.sprite) {
     const weaponModel = <HTMLImageElement>document.querySelector(".sprites ." + playerModel.weapon.sprite);
@@ -898,6 +901,7 @@ function renderPlayerPortrait() {
 
 function renderPlayerModel(size: number, canvas: HTMLCanvasElement, ctx: any) {
   canvas.width = canvas.width; // Clear canvas
+  const sex = player.sex === "male" ? "" : capitalizeFirstLetter(player.sex);
   if (player.isDead) return;
   const bodyModel = <HTMLImageElement>document.querySelector(".sprites ." + player.race + "Model" + capitalizeFirstLetter(player.sex));
   const earModel = <HTMLImageElement>document.querySelector(".sprites ." + player.race + "Ears");
@@ -917,19 +921,19 @@ function renderPlayerModel(size: number, canvas: HTMLCanvasElement, ctx: any) {
   ctx?.drawImage(faceModel, baseCanvas.width / 2 - size / 2, baseCanvas.height / 2 - size / 2, size, size);
   if (!player.helmet?.coversHair) ctx?.drawImage(hairModel, baseCanvas.width / 2 - size / 2, baseCanvas.height / 2 - size / 2, size, size);
   if (player.helmet?.sprite) {
-    const helmetModel = <HTMLImageElement>document.querySelector(".sprites ." + player.helmet.sprite);
+    const helmetModel = <HTMLImageElement>document.querySelector(".sprites ." + player.helmet.sprite + sex);
     ctx?.drawImage(helmetModel, baseCanvas.width / 2 - size / 2, baseCanvas.height / 2 - size / 2, size, size);
   }
   if (player.gloves?.sprite) {
-    const glovesModel = <HTMLImageElement>document.querySelector(".sprites ." + player.gloves.sprite);
+    const glovesModel = <HTMLImageElement>document.querySelector(".sprites ." + player.gloves.sprite + sex);
     ctx?.drawImage(glovesModel, baseCanvas.width / 2 - size / 2, baseCanvas.height / 2 - size / 2, size, size);
   }
   if (player.boots?.sprite) {
-    const bootsModel = <HTMLImageElement>document.querySelector(".sprites ." + player.boots.sprite);
+    const bootsModel = <HTMLImageElement>document.querySelector(".sprites ." + player.boots.sprite + sex);
     ctx?.drawImage(bootsModel, baseCanvas.width / 2 - size / 2, baseCanvas.height / 2 - size / 2, size, size);
   }
   if (player.legs?.sprite) {
-    const leggingsModel = <HTMLImageElement>document.querySelector(".sprites ." + player.legs.sprite);
+    const leggingsModel = <HTMLImageElement>document.querySelector(".sprites ." + player.legs.sprite + sex);
     ctx?.drawImage(leggingsModel, baseCanvas.width / 2 - size / 2, baseCanvas.height / 2 - size / 2, size, size);
   }
   else if (!player.legs?.sprite) {
@@ -937,7 +941,7 @@ function renderPlayerModel(size: number, canvas: HTMLCanvasElement, ctx: any) {
     ctx?.drawImage(leggings, baseCanvas.width / 2 - size / 2, baseCanvas.height / 2 - size / 2, size, size);
   }
   if (player.chest?.sprite) {
-    const chestModel = <HTMLImageElement>document.querySelector(".sprites ." + player.chest.sprite);
+    const chestModel = <HTMLImageElement>document.querySelector(".sprites ." + player.chest.sprite + sex);
     ctx?.drawImage(chestModel, baseCanvas.width / 2 - size / 2, baseCanvas.height / 2 - size / 2, size, size);
   }
   if (player.weapon?.sprite) {